% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{add_font}
\alias{add_font}
\title{Register a font for graphics}
\usage{
add_font(
  font = c("Arial", "Open Sans", "Roboto", "Fira Code", "Fira Sans Condensed")
)
}
\arguments{
\item{font}{one of "Arial", "Open Sans", "Roboto", "Fira Code" or "Fira Sans Condensed".}
}
\value{
TRUE if the operation went ok.
}
\description{
Register with package 'systemfonts' a font
from the package so that the font is available
when the graphics are built.
}
\examples{
add_font(font = "Open Sans")
add_font(font = "Roboto")
}
