% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{addFontHtmlDependency}
\alias{addFontHtmlDependency}
\title{Use a font in Shiny or Markdown}
\usage{
addFontHtmlDependency(
  font = c("Arial", "Open Sans", "Roboto", "Fira Code", "Fira Sans Condensed")
)
}
\arguments{
\item{font}{one of "Arial", "Open Sans", "Roboto", "Fira Code" or "Fira Sans Condensed".}
}
\value{
an HTML object
}
\description{
Add an empty \verb{<style>} HTML element attached
to an \href{htmlDependency}{HTML Dependency} containing
the css and the font files so that the font is available
in the HTML page.

The htmlDependency is defined with function \code{\link[=fontHtmlDependency]{fontHtmlDependency()}}.
}
\examples{
addFontHtmlDependency(font = "Open Sans")
addFontHtmlDependency(font = "Roboto")
}
