% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{fontHtmlDependency}
\alias{fontHtmlDependency}
\title{Font HTML dependency}
\usage{
fontHtmlDependency(
  font = c("Arial", "Open Sans", "Roboto", "Fira Code", "Fira Sans Condensed")
)
}
\arguments{
\item{font}{one of "Arial", "Open Sans", "Roboto", "Fira Code" or "Fira Sans Condensed".}
}
\value{
an object defined with \code{\link[=htmlDependency]{htmlDependency()}}.
}
\description{
Create an HTML dependency ready
to be used in 'Shiny' or 'R Markdown'. It allows
to use specific fonts in HTML page.
}
