% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{install_font_str}
\alias{install_font_str}
\title{Shell command to install a font}
\usage{
install_font_str(
  font = c("Arial", "Open Sans", "Roboto", "Fira Code", "Fira Sans Condensed"),
  dir = "custom-fonts",
  platform = "ubuntu"
)
}
\arguments{
\item{font}{one of "Arial", "Open Sans", "Roboto", "Fira Code" or "Fira Sans Condensed".}

\item{dir}{the name of the sub-directory to use when copying the font files.}

\item{platform}{not yet supported, only ubuntu is supported for now.}
}
\value{
the shell command as a string
}
\description{
Create a string containing the shell command to execute
so that the font is installed on the system. Its execution should require root
permissions.
}
\examples{
install_font_str(font = "Open Sans")
}
